bug fix with password
