+++ backend/auth_service/app/services/user_service.py
@@ -8,12 +8,9 @@ from core.httpexceptions import (
     UserAlreadyExistsException,
     EmailSendException
 )
-from utils.password_utils import validate_password, generate_new_password
+from utils.password_utils import validate_password, generate_new_password, get_password_hash
 from utils.email_service import send_email_with_new_password
 logger = logging.getLogger(__name__)
-
-from utils.password_utils import validate_password, generate_new_password
-
 
 class UserService:
     @classmethod
@@ -73,28 +70,6 @@ class UserService:
 
         logger.info(f"User data for JWT retrieved for email: {email}")
         return user_data
-
-    @classmethod
-    async def change_password(cls, email: EmailStr, old_password: str, new_password: str, confirm_password: str) -> dict:
-        """
-        Change a user's password after validating the old password.
-        """
-        if new_password != confirm_password:
-            raise InvalidPasswordException(email=email, detail="New password and confirmation do not match.")
-
-        # Validate old password
-        user = await UserRepository.find_user_by_email(email)
-        if not user:
-            raise UserNotFoundException(email=email)
-
-        password_from_db = await UserRepository.find_password_by_email(email)
-        if not validate_password(old_password, password_from_db):
-            raise InvalidPasswordException(email=email, detail="Old password is incorrect.")
-
-        # Update to new password
-        await UserRepository.update_password(email=email, new_password=new_password)
-        logger.info(f"Password changed successfully for user {email}.")
-        return {"message": "Password changed successfully."}
 
     @classmethod
     async def change_password(cls, email: EmailStr, old_password: str, new_password: str, confirm_password: str) -> dict:
